+++ packages/app-kit/src/genre-tree/models/TrackList.ts
@@ -1,21 +1,21 @@
-import { UploadedTrackDetailed } from "../schemas/uploaded-track/detailed";
+import { TrackDetailed } from "../schemas/track/detailed";
 import TrackListOrigin, {
   TrackListOriginFromUploadedTrack,
   TrackListOriginFromCriteriaPlaylist,
 } from "./TrackListOrigin";
 
 export default class TrackList {
-  constructor(public uploadedTracks: UploadedTrackDetailed[], public origin: TrackListOrigin) {}
+  constructor(public uploadedTracks: TrackDetailed[], public origin: TrackListOrigin) {}
 }
 
 export class TrackListFromUploadedTrack extends TrackList {
-  constructor(public uploadedTracks: UploadedTrackDetailed[], public origin: TrackListOriginFromUploadedTrack) {
+  constructor(public uploadedTracks: TrackDetailed[], public origin: TrackListOriginFromUploadedTrack) {
     super(uploadedTracks, origin);
   }
 }
 
 export class TrackListFromCriteriaPlaylist extends TrackList {
-  constructor(public uploadedTracks: UploadedTrackDetailed[], public origin: TrackListOriginFromCriteriaPlaylist) {
+  constructor(public uploadedTracks: TrackDetailed[], public origin: TrackListOriginFromCriteriaPlaylist) {
     super(uploadedTracks, origin);
   }
 }