+++ packages/app-kit/src/genre-tree/schemas/criteria-playlist/detailed.ts
@@ -1,17 +1,17 @@
 import { z } from "zod";
 
 import { UuidResourceSchema } from "../uuid-resource";
-import { UploadedTrackPlaylistRelWithoutPlaylistSchema } from "../uploaded-track-playlist-rel/without-playlist";
+import { TrackPlaylistRelWithoutPlaylistSchema } from "../uploaded-track-playlist-rel/without-playlist";
 import { CriteriaMinimumSchema } from "../criteria/minimum";
 import { CriteriaPlaylistMinimumSchema } from "./minimum";
 
 export const CriteriaPlaylistDetailedSchema = UuidResourceSchema.extend({
   name: z.string(),
-  uploadedTrackPlaylistRelations: z.array(UploadedTrackPlaylistRelWithoutPlaylistSchema),
-  uploadedTracksCount: z.number(),
+  trackPlaylistRelations: z.array(TrackPlaylistRelWithoutPlaylistSchema),
+  tracksCount: z.number(),
   durationInSec: z.number(),
   durationStrInHourMinSec: z.string(),
-  uploadedTracksArchivedCount: z.number(),
+  tracksArchivedCount: z.number(),
   criteria: CriteriaMinimumSchema,
   parent: CriteriaPlaylistMinimumSchema.nullable(),
   root: CriteriaPlaylistMinimumSchema,